% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CholWishart-package.r
\docType{package}
\name{CholWishart-package}
\alias{CholWishart}
\alias{CholWishart-package}
\title{Cholesky Factor of a Wishart or Inverse Wishart}
\description{
The most common use for this package is likely fast sampling from the
inverse Wishart or use of the multivariate gamma and digamma functions.
This is a package for fast computation of various functions related to
the Wishart distribution, such as sampling from the Cholesky factorization
of the Wishart, sampling from the inverse Wishart,
sampling from the Cholesky factorization of the inverse Wishart,
sampling from the pseudo Wishart,
computing densities for the Wishart and inverse Wishart,
and computing a few auxiliary functions
such as the multivariate gamma and digamma functions.
Many of these functions are written in C to maximize efficiency.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gzt/CholWishart https://gzt.github.io/CholWishart/}
  \item Report bugs at \url{https://github.com/gzt/CholWishart/issues}
}

}
\author{
\strong{Maintainer}: Geoffrey Thompson \email{gzthompson@gmail.com} (\href{https://orcid.org/0000-0003-2436-8822}{ORCID})

Other contributors:
\itemize{
  \item R Core Team [contributor]
}

}
\keyword{internal}
