% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CholWishart-package.r
\docType{package}
\name{CholWishart}
\alias{CholWishart}
\alias{CholWishart-package}
\title{CholWishart.}
\description{
A package for fast computation various functions related to the Wishart distribution, such as sampling from the Cholesky factor of the Wishart, sampling from the inverse Wishart, sampling from the Cholesky factor of the inverse Wishart, computing densities for the Wishart and inverse Wishart, and computing a few auxiliart functions such as the multivariate gamma and digamma functions. Many of these functions are written in C to maximize efficiency.
}
