% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CholWishart-package.r
\docType{package}
\name{CholWishart}
\alias{CholWishart}
\alias{CholWishart-package}
\title{Cholesky Factor of a Wishart or Inverse Wishart}
\description{
A package for fast computation of various functions related to the Wishart distribution, such as sampling from the Cholesky factor of the Wishart, sampling from the inverse Wishart, sampling from the Cholesky factor of the inverse Wishart, computing densities for the Wishart and inverse Wishart, and computing the multivariate gamma and digamma functions. Many of these functions are written in C to maximize speed.
}
