% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wishart.R
\name{mvdigamma}
\alias{mvdigamma}
\title{Multivariate Digamma Function}
\usage{
mvdigamma(x, p)
}
\arguments{
\item{x}{non-negative numeric vector, matrix, or array}

\item{p}{positive integer, dimension of a square matrix}
}
\value{
vector of values of multivariate digamma function.
}
\description{
A special mathematical function related to the gamma function,
    generalized for multivariate distributions.
    The multivariate digamma function is the derivative of the multivariate gamma function
    for for p = 1 is the same as the univariate digamma function.
}
\examples{
digamma(1:10)
mvdigamma(1:10,1)
}
\references{
A. K. Gupta and D. K. Nagar 1999. \emph{Matrix variate distributions}. Chapman and Hall.
}
