% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wishart.R
\name{mvdigamma}
\alias{mvdigamma}
\title{Multivariate Digamma}
\usage{
mvdigamma(x, p)
}
\arguments{
\item{x}{non-negative numeric vector, matrix, or array}

\item{p}{positive integer, dimension of a square matrix}
}
\value{
vector of values of multivariate digamma function.
}
\description{
A special mathematical function related to the gamma function,
    generalized for multivariate distributions.
    The digamma is the derivative of the gamma.
}
\examples{
digamma(1:10)
mvdigamma(1:10,1)
}
